feat: private shops (#7)
* build: update scripts

* chore: update seeder

* refactor: move user-menu into users feature

* feat: shared checkbox-input

* fix: shared search-input with label and placeholder props

* fix: disable secure cookie for dev

* feat: shared tooltip

* fix: sign in remember me hint

* fix: users search

* fix: handleErrors to accept status

* feat: shared password input

* fix: hide pagination when pagination has no value

* feat: utils type

* fix: meta query type

* feat: shared url link input

* fix: authentication components and validation

* feat: shop card component

* fix: shared modal responsive style

* feat: shop schema private key

* fix: shop collection to accept collection path with default

* feat: users shop collection

* fix: shop update and delete implementations

- include for created details on update
- check for user roles on delete

* feat: private shop parameter for create shop

* feat: private shop read, update and delete endpoint

* fix: shop component and validation

* fix: private shop route

* fix: type import

* fix: show banner when unauthenticated

* fix: base api to include only approved shops

* chore: remove moved seeder

* feat: private shop nav

* fix: hide search webkit style

* chore: update env, linter, gitignore and prettier

* ci: remove deployment step

* fix: function entrypoint

* build: sveltekit config

* build: update deps

* chore: format code

+++ src/routes/shops/[uid]/index.ts
@@ -51,8 +51,6 @@ export const patch: RequestHandler<PatchOutput | ErrorResponse> = async ({
 
         const collection = new ShopCollection();
 
-        const shopData = await collection.get(params.uid);
-
         await collection.set({
             uid: params.uid,
             name: payload.name,
@@ -67,17 +65,12 @@ export const patch: RequestHandler<PatchOutput | ErrorResponse> = async ({
                 country: payload.country,
             },
             status: payload.status,
-            createdAt: shopData.createdAt,
-            createdBy: {
-                uid: shopData.createdBy.uid,
-                name: shopData.createdBy.name,
-            },
         });
 
         return {
             status: 200,
             body: {
-                message: 'Shop "' + payload.name + '" has been saved.',
+                message: payload.name + " shop has been saved.",
             },
         };
     } catch (error) {
@@ -89,7 +82,19 @@ export type DeleteOutput = {
     message: string;
 };
 
-export const del: RequestHandler<DeleteOutput> = async ({ params }) => {
+export const del: RequestHandler<DeleteOutput | ErrorResponse> = async ({ locals, params }) => {
+    const user = locals.user;
+
+    if (!user) {
+        return handleError(new Error("Not authenticated."));
+    }
+
+    const userData = user.data();
+
+    if (!userData.customClaims.isAdmin) {
+        return handleError(new Error("Not authorized."));
+    }
+
     const collection = new ShopCollection();
     await collection.delete(params.uid);
